+++ packages/admin/resources/lang/en/partials.php
@@ -6,6 +6,16 @@ return [
      */
     'activity-log.added_images.description' => 'Added :count images',
     /**
+     * Associations.
+     */
+    'products.associations.heading' => 'Associations',
+    'products.associations.cross-sell' => 'Cross Sell',
+    'products.associations.up-sell' => 'Up Sell',
+    'products.associations.alternate' => 'Alternate',
+    'products.associations.show_inverse' => 'Show inverse',
+    'products.associations.add_inverse' => 'Add inverse association',
+    'products.associations.add_association' => 'Add association',
+    /**
      * Availability.
      */
     'availability.heading'                                               => 'Availability',