+++ packages/admin/resources/views/components/modal/dialog.blade.php
@@ -18,4 +18,4 @@
     @if($form)
       </form>
     @endif
-</x-hub::>
+</x-hub::modal>