+++ packages/admin/src/Http/Livewire/Components/Products/AbstractProduct.php
@@ -16,6 +16,7 @@ use GetCandy\Hub\Jobs\Products\GenerateVariants;
 use GetCandy\Models\AttributeGroup;
 use GetCandy\Models\Collection as ModelsCollection;
 use GetCandy\Models\Product;
+use GetCandy\Models\ProductAssociation;
 use GetCandy\Models\ProductOption;
 use GetCandy\Models\ProductType;
 use GetCandy\Models\ProductVariant;
@@ -124,6 +125,34 @@ abstract class AbstractProduct extends Component
      */
     public $variantAttributes;
 
+    /**
+     * Whether to show inverse associations.
+     *
+     * @var bool
+     */
+    public $showInverseAssociations = false;
+
+    /**
+     * The base association type to use.
+     *
+     * @var string
+     */
+    public $associationType = 'cross-sell';
+
+    /**
+     * The current product associations.
+     *
+     * @var Collection
+     */
+    public Collection $associations;
+
+    /**
+     * Associations that need removing.
+     *
+     * @var array
+     */
+    public array $associationsToRemove = [];
+
     protected function getListeners()
     {
         return array_merge([
@@ -131,6 +160,7 @@ abstract class AbstractProduct extends Component
             'productOptionCreated'          => 'resetOptionView',
             'option-manager.selectedValues' => 'setOptionValues',
             'urlSaved'                      => 'refreshUrls',
+            'product-search.selected'       => 'updateAssociations',
             'collectionSearch.selected'     => 'selectCollections',
         ], $this->getHasImagesListeners());
     }
@@ -163,6 +193,7 @@ abstract class AbstractProduct extends Component
             'urls'                    => 'array',
             'collections'             => 'nullable|array',
             'variant.tax_ref'         => 'nullable|string|max:255',
+            'associations.*.type'     => 'required|string',
             'variant.sku'             => get_validation('products', 'sku', [
                 'alpha_dash',
                 'max:255',
@@ -349,6 +380,26 @@ abstract class AbstractProduct extends Component
 
             $this->product->channels()->sync($channels);
 
+            if (count($this->associationsToRemove)) {
+                ProductAssociation::whereIn('id', $this->associationsToRemove)->delete();
+            }
+
+            $this->associations->each(function ($assoc) {
+                if (! empty($assoc['id'])) {
+                    ProductAssociation::find($assoc['id'])->update([
+                        'type' => $assoc['type'],
+                    ]);
+
+                    return;
+                }
+
+                ProductAssociation::create([
+                    'product_target_id' => $assoc['inverse'] ? $this->product->id : $assoc['target_id'],
+                    'product_parent_id' => $assoc['inverse'] ? $assoc['target_id'] : $this->product->id,
+                    'type' => $assoc['type'],
+                ]);
+            });
+
             $this->product->collections()->detach(
                 $this->collectionsToDetach->pluck('id')
             );
@@ -528,6 +579,93 @@ abstract class AbstractProduct extends Component
     }
 
     /**
+     * Sync initial product associations.
+     *
+     * @return void
+     */
+    public function syncAssociations()
+    {
+        $this->associations = $this->product->associations
+            ->merge($this->product->inverseAssociations)
+            ->map(function ($assoc) {
+                $inverse = $assoc->target->id == $this->product->id;
+
+                $product = $inverse ? $assoc->parent : $assoc->target;
+
+                return [
+                    'id' => $assoc->id,
+                    'inverse' => $inverse,
+                    'target_id' => $product->id,
+                    'thumbnail' => optional($product->thumbnail)->getUrl('small'),
+                    'name' => $product->translateAttribute('name'),
+                    'type' => $assoc->type,
+                ];
+            });
+    }
+
+    /**
+     * Update the associations.
+     *
+     * @param  array  $selectedIds
+     * @return void
+     */
+    public function updateAssociations($selectedIds)
+    {
+        $selectedProducts = Product::findMany($selectedIds)->map(function ($product) {
+            return [
+                'inverse' => (bool) $this->showInverseAssociations,
+                'target_id' => $product->id,
+                'thumbnail' => optional($product->thumbnail)->getUrl('small'),
+                'name' => $product->translateAttribute('name'),
+                'type' => $this->associationType,
+            ];
+        });
+
+        $this->associations = $this->associations->count() ?
+            $this->associations->merge($selectedProducts) :
+            $selectedProducts;
+
+        $this->emit('updatedExistingProductAssociations', $this->associatedProductIds);
+    }
+
+    /**
+     * Open the association browser with a given type.
+     *
+     * @param  string  $type
+     * @return void
+     */
+    public function openAssociationBrowser($type)
+    {
+        $this->associationType = $type;
+        $this->emit('showBrowser', 'product-associations');
+    }
+
+    /**
+     * Remove an association.
+     *
+     * @param  int  $index
+     * @return void
+     */
+    public function removeAssociation($index)
+    {
+        $this->associationsToRemove[] = $this->associations[$index]['id'];
+
+        $this->associations->forget($index);
+    }
+
+    /**
+     * The associated product ids.
+     *
+     * @return void
+     */
+    public function getAssociatedProductIdsProperty()
+    {
+        return collect(
+            $this->associations->map(fn ($assoc) => ['id' => $assoc['target_id']])
+        );
+    }
+
+    /**
      * Returns the attribute data.
      *
      * @return array
@@ -664,6 +802,13 @@ abstract class AbstractProduct extends Component
                 'title'      => __('adminhub::menu.product.shipping'),
                 'id'         => 'shipping',
                 'hidden'     => $this->getVariantsCount() > 1,
+                'has_errors' => $this->errorBag->hasAny([
+                ]),
+            ],
+            [
+                'title'      => __('adminhub::menu.product.associations'),
+                'id'         => 'associations',
+                'hidden'     => false,
                 'has_errors' => $this->errorBag->hasAny([
                 ]),
             ],