+++ packages/admin/src/Http/Livewire/Components/Products/ProductCreate.php
@@ -39,6 +39,7 @@ class ProductCreate extends AbstractProduct
         );
 
         $this->syncAvailability();
+        $this->syncAssociations();
         $this->syncCollections();
     }